#4023 Moved waiver model from dissemination module into audit module and updated waiver_types field
